pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=61.2",
  "setuptools-scm>=8",
]

[project]
name = "contributors-txt"
description = "Generate a list of contributors automatically from git data"
readme.content-type = "text/markdown"
readme.file = "README.md"
authors = [
  { name = "Pierre Sassoulas", email = "pierre.sassoulas@gmail.com" },
]
requires-python = ">=3.9.0"
classifiers = [
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dynamic = [ "version" ]
dependencies = []
optional-dependencies.test = [
  "coverage<8",
  "coveralls",
  "pre-commit>=1.16",
  "pylint==4.0.4",
  "pytest-cov",
  "pytest-remaster",
  "pytest-vcr",
]
urls.Homepage = "https://github.com/Pierre-Sassoulas/contributors-txt"
scripts.contributors-txt = "contributors_txt.__main__:main"
scripts.contributors-txt-extract-comment = "contributors_txt.extract_comment:main"
scripts.contributors-txt-normalize-configuration = "contributors_txt.normalize:main"

[tool.setuptools]
include-package-data = false
packages.find.include = [
  "contributors_txt*",
]
packages.find.exclude = [
  "tests.*",
]
packages.find.namespaces = false

[tool.setuptools_scm]

[tool.ruff]
target-version = "py39"
line-length = 88
src = [
  "src",
]
format.docstring-code-format = true
lint.select = [
  "B",       # bugbear
  "E",       # pycodestyle
  "F",       # pyflakes
  "FA100",   # add future annotations
  "I",       # isort
  "PGH004",  # pygrep-hooks - Use specific rule codes when using noqa
  "PIE",     # flake8-pie
  "PLC",     # pylint convention
  "PLE",     # pylint error
  "PLR",     # pylint refactor
  "PLR1714", # Consider merging multiple comparisons
  "PLW",     # pylint warning
  "PYI",     # flake8-pyi
  "RUF",     # ruff
  "T100",    # flake8-debugger
  "UP",      # pyupgrade
  "W",       # pycodestyle
]
lint.ignore = [
  "PLR2004",
]

[tool.pylint]
main.disable = [ "missing-docstring", "too-few-public-methods", "wrong-import-order" ]

[tool.mypy]
scripts_are_modules = true
no_implicit_optional = true
warn_redundant_casts = true
show_error_codes = true
strict = true
overrides = [
  { module = [
    "setuptools",
  ], ignore_missing_imports = true },
  { module = [
    "pytest",
  ], ignore_missing_imports = true },
]

[tool.pytest]
ini_options.remaster-by-default = false
